fix: myPage에서 userId없을 떄 api요청 안하게 수정
